refactor(engine): keep default rule severities in the registry
