+++ src/engine/rules/contraction.ts
@@ -1,5 +1,6 @@
 import { scanLines } from "../scan.ts"
 import type { Violation } from "../types.ts"
+import { DEFAULT_SEVERITIES } from "./registry.ts"
 
 const CONTRACTION = /\b\w+['’](?:t|re|ve|ll|d|m)\b/gi
 
@@ -11,7 +12,7 @@ const CONTRACTION_S =
 export function contraction(lines: readonly string[]): Violation[] {
   return [...scanLines(lines, CONTRACTION), ...scanLines(lines, CONTRACTION_S)].map((match) => ({
     ruleId: "contraction",
-    severity: "hard" as const,
+    severity: DEFAULT_SEVERITIES.contraction,
     message: `Do not use a contraction. Write the words in full. Found "${match.found}".`,
     line: match.line,
     column: match.column,
