+++ src/engine/rules/sentence-length.ts
@@ -1,5 +1,6 @@
 import type { Sentence } from "../sentences.ts"
 import type { Violation } from "../types.ts"
+import { DEFAULT_SEVERITIES } from "./registry.ts"
 
 const countWords = (text: string): number => text.split(/\s+/).filter((word) => word !== "").length
 
@@ -8,7 +9,7 @@ export function sentenceLength(sentence: Sentence, maxWords: number): Violation 
   if (count <= maxWords) return undefined
   return {
     ruleId: "sentence-length",
-    severity: "hard",
+    severity: DEFAULT_SEVERITIES["sentence-length"],
     message: `Sentence has ${count} words; the maximum is ${maxWords}.`,
     line: sentence.line,
     column: sentence.column,
