+++ src/engine/suppression.ts
@@ -2,7 +2,7 @@ import type { LineCommentSpan } from "./comments.ts"
 import type { ScopedViolation } from "./diff-match.ts"
 import { htmlComments } from "./html.ts"
 import { type MarkdownHtmlComment, markdownHtmlComments } from "./markdown.ts"
-import { type RuleId, ruleIds } from "./rules/registry.ts"
+import { DEFAULT_SEVERITIES, type RuleId, ruleIds } from "./rules/registry.ts"
 import { lineOffsets } from "./scan.ts"
 import type { LintKind } from "./types.ts"
 
@@ -101,7 +101,7 @@ const invalidFinding = (
   return {
     violation: {
       ruleId: "invalid-suppression",
-      severity: "hard",
+      severity: DEFAULT_SEVERITIES["invalid-suppression"],
       message,
       line: directive.line,
       column: directive.column,
